refactor, move get-cloud-user from snapping to legacy
