+++ scopes/component/snapping/esm.mjs
@@ -2,8 +2,5 @@
 import cjsModule from './index.js';
 
 export const SnappingAspect = cjsModule.SnappingAspect;
-export const getBitCloudUsername = cjsModule.getBitCloudUsername;
-export const getBitCloudUser = cjsModule.getBitCloudUser;
-export const getBasicLog = cjsModule.getBasicLog;
 
 export default cjsModule;
