+++ scopes/component/snapping/index.ts
@@ -1,7 +1,6 @@
 import { SnappingAspect } from './snapping.aspect';
 
 export type { BitCloudUser } from './tag-model-component';
-export { getBitCloudUsername, getBitCloudUser, getBasicLog } from './tag-model-component';
 export type { SnappingMain, TagResults, SnapResults } from './snapping.main.runtime';
 export default SnappingAspect;
 export { SnappingAspect };
