+++ scopes/scope/sign/sign.main.runtime.ts
@@ -6,8 +6,9 @@ import { ScopeAspect, ScopeMain } from '@teambit/scope';
 import { BuilderAspect, BuilderMain } from '@teambit/builder';
 import { isSnap } from '@teambit/component-version';
 import { Component, ComponentID } from '@teambit/component';
-import { getBasicLog, SnappingAspect, SnappingMain } from '@teambit/snapping';
+import { SnappingAspect, SnappingMain } from '@teambit/snapping';
 import ConsumerComponent from '@teambit/legacy/dist/consumer/component';
+import { getBasicLog } from '@teambit/legacy/dist/utils/bit/basic-log';
 import { BuildStatus, CENTRAL_BIT_HUB_URL, CENTRAL_BIT_HUB_NAME } from '@teambit/legacy/dist/constants';
 import { getScopeRemotes } from '@teambit/legacy/dist/scope/scope-remotes';
 import { PostSign } from '@teambit/legacy/dist/scope/actions';
